brain-tec assignment/rodriguez_cortes_jose_carlos/requirements.txt
flask==1.1.2
mysql.connector.python
pytest
coverage
defusedxml